Revert "Range check optimizations"
