+++ .gitignore
@@ -15,4 +15,3 @@
 # vendor/
 
 gnark-ed25519
-gnark.pprof
