+++ challenger/challenger.go
@@ -8,7 +8,6 @@ import (
 	gl "github.com/succinctlabs/gnark-plonky2-verifier/goldilocks"
 	"github.com/succinctlabs/gnark-plonky2-verifier/poseidon"
 	"github.com/succinctlabs/gnark-plonky2-verifier/types"
-	"github.com/succinctlabs/gnark-plonky2-verifier/variables"
 )
 
 type Chip struct {
@@ -114,12 +113,12 @@ func (c *Chip) GetHash() poseidon.GoldilocksHashOut {
 }
 
 func (c *Chip) GetFriChallenges(
-	commitPhaseMerkleCaps []variables.FriMerkleCap,
-	finalPoly variables.PolynomialCoeffs,
+	commitPhaseMerkleCaps []types.FriMerkleCap,
+	finalPoly types.PolynomialCoeffs,
 	powWitness gl.Variable,
 	degreeBits uint64,
 	config types.FriConfig,
-) variables.FriChallenges {
+) types.FriChallenges {
 	numFriQueries := config.NumQueryRounds
 	friAlpha := c.GetExtensionChallenge()
 
@@ -135,7 +134,7 @@ func (c *Chip) GetFriChallenges(
 	friPowResponse := c.GetChallenge()
 	friQueryIndices := c.GetNChallenges(numFriQueries)
 
-	return variables.FriChallenges{
+	return types.FriChallenges{
 		FriAlpha:        friAlpha,
 		FriBetas:        friBetas,
 		FriPowResponse:  friPowResponse,
@@ -153,7 +152,7 @@ func (c *Chip) duplexing() {
 		panic("something went wrong")
 	}
 
-	glApi := gl.New(c.api)
+	glApi := gl.NewChip(c.api)
 
 	for i := 0; i < len(c.inputBuffer); i++ {
 		c.spongeState[i] = glApi.Reduce(c.inputBuffer[i])
