+++ fri/fri.go
@@ -11,65 +11,26 @@ import (
 	gl "github.com/succinctlabs/gnark-plonky2-verifier/goldilocks"
 	"github.com/succinctlabs/gnark-plonky2-verifier/poseidon"
 	"github.com/succinctlabs/gnark-plonky2-verifier/types"
-	"github.com/succinctlabs/gnark-plonky2-verifier/variables"
 )
 
 type Chip struct {
-	api               frontend.API             `gnark:"-"`
-	gl                gl.Chip                  `gnark:"-"`
-	poseidonBN254Chip *poseidon.BN254Chip      `gnark:"-"`
-	commonData        *types.CommonCircuitData `gnark:"-"`
-	friParams         *types.FriParams         `gnark:"-"`
+	api               frontend.API `gnark:"-"`
+	gl                gl.Chip      `gnark:"-"`
+	poseidonBN254Chip *poseidon.BN254Chip
+	friParams         *types.FriParams `gnark:"-"`
 }
 
 func NewChip(
 	api frontend.API,
-	commonData *types.CommonCircuitData,
 	friParams *types.FriParams,
 ) *Chip {
 	poseidonBN254Chip := poseidon.NewBN254Chip(api)
 	return &Chip{
 		api:               api,
 		poseidonBN254Chip: poseidonBN254Chip,
-		commonData:        commonData,
 		friParams:         friParams,
-		gl:                *gl.New(api),
+		gl:                *gl.NewChip(api),
 	}
-}
-
-func (f *Chip) GetInstance(zeta gl.QuadraticExtensionVariable) InstanceInfo {
-	zetaBatch := BatchInfo{
-		Point:       zeta,
-		Polynomials: friAllPolys(f.commonData),
-	}
-
-	g := gl.PrimitiveRootOfUnity(f.commonData.DegreeBits)
-	zetaNext := f.gl.MulExtension(
-		gl.NewVariable(g.Uint64()).ToQuadraticExtension(),
-		zeta,
-	)
-
-	zetaNextBath := BatchInfo{
-		Point:       zetaNext,
-		Polynomials: friZSPolys(f.commonData),
-	}
-
-	return InstanceInfo{
-		Oracles: friOracles(f.commonData),
-		Batches: []BatchInfo{zetaBatch, zetaNextBath},
-	}
-}
-
-func (f *Chip) ToOpenings(c variables.OpeningSet) Openings {
-	values := c.Constants                         // num_constants + 1
-	values = append(values, c.PlonkSigmas...)     // num_routed_wires
-	values = append(values, c.Wires...)           // num_wires
-	values = append(values, c.PlonkZs...)         // num_challenges
-	values = append(values, c.PartialProducts...) // num_challenges * num_partial_products
-	values = append(values, c.QuotientPolys...)   // num_challenges * quotient_degree_factor
-	zetaBatch := OpeningBatch{Values: values}
-	zetaNextBatch := OpeningBatch{Values: c.PlonkZsNext}
-	return Openings{Batches: []OpeningBatch{zetaBatch, zetaNextBatch}}
 }
 
 func (f *Chip) assertLeadingZeros(powWitness gl.Variable, friConfig types.FriConfig) {
@@ -100,8 +61,8 @@ func (f *Chip) verifyMerkleProofToCapWithCapIndex(
 	leafData []gl.Variable,
 	leafIndexBits []frontend.Variable,
 	capIndexBits []frontend.Variable,
-	merkleCap variables.FriMerkleCap,
-	proof *variables.FriMerkleProof,
+	merkleCap types.FriMerkleCap,
+	proof *types.FriMerkleProof,
 ) {
 	currentDigest := f.poseidonBN254Chip.HashOrNoop(leafData)
 	for i, sibling := range proof.Siblings {
@@ -139,7 +100,7 @@ func (f *Chip) verifyMerkleProofToCapWithCapIndex(
 	f.api.AssertIsEqual(currentDigest, merkleCapEntry)
 }
 
-func (f *Chip) verifyInitialProof(xIndexBits []frontend.Variable, proof *variables.FriInitialTreeProof, initialMerkleCaps []variables.FriMerkleCap, capIndexBits []frontend.Variable) {
+func (f *Chip) verifyInitialProof(xIndexBits []frontend.Variable, proof *types.FriInitialTreeProof, initialMerkleCaps []types.FriMerkleCap, capIndexBits []frontend.Variable) {
 	if len(proof.EvalsProofs) != len(initialMerkleCaps) {
 		panic("length of eval proofs in fri proof should equal length of initial merkle caps")
 	}
@@ -226,7 +187,7 @@ func (f *Chip) calculateSubgroupX(
 
 func (f *Chip) friCombineInitial(
 	instance InstanceInfo,
-	proof variables.FriInitialTreeProof,
+	proof types.FriInitialTreeProof,
 	friAlpha gl.QuadraticExtensionVariable,
 	subgroupX_QE gl.QuadraticExtensionVariable,
 	precomputedReducedEval []gl.QuadraticExtensionVariable,
@@ -267,7 +228,7 @@ func (f *Chip) friCombineInitial(
 	return sum
 }
 
-func (f *Chip) finalPolyEval(finalPoly variables.PolynomialCoeffs, point gl.QuadraticExtensionVariable) gl.QuadraticExtensionVariable {
+func (f *Chip) finalPolyEval(finalPoly types.PolynomialCoeffs, point gl.QuadraticExtensionVariable) gl.QuadraticExtensionVariable {
 	ret := gl.ZeroExtension()
 	for i := len(finalPoly.Coeffs) - 1; i >= 0; i-- {
 		ret = f.gl.MulAddExtension(ret, point, finalPoly.Coeffs[i])
@@ -394,14 +355,14 @@ func (f *Chip) computeEvaluation(
 
 func (f *Chip) verifyQueryRound(
 	instance InstanceInfo,
-	challenges *variables.FriChallenges,
+	challenges *types.FriChallenges,
 	precomputedReducedEval []gl.QuadraticExtensionVariable,
-	initialMerkleCaps []variables.FriMerkleCap,
-	proof *variables.FriProof,
+	initialMerkleCaps []types.FriMerkleCap,
+	proof *types.FriProof,
 	xIndex gl.Variable,
 	n uint64,
 	nLog uint64,
-	roundProof *variables.FriQueryRound,
+	roundProof *types.FriQueryRound,
 ) {
 	f.assertNoncanonicalIndicesOK()
 	xIndex = f.gl.Reduce(xIndex)
@@ -507,9 +468,9 @@ func (f *Chip) verifyQueryRound(
 func (f *Chip) VerifyFriProof(
 	instance InstanceInfo,
 	openings Openings,
-	friChallenges *variables.FriChallenges,
-	initialMerkleCaps []variables.FriMerkleCap,
-	friProof *variables.FriProof,
+	friChallenges *types.FriChallenges,
+	initialMerkleCaps []types.FriMerkleCap,
+	friProof *types.FriProof,
 ) {
 	// TODO:  Check fri config
 	/* if let Some(max_arity_bits) = params.max_arity_bits() {
