--- goldilocks/.gitignore
@@ -1,1 +1,0 @@
-gnark.pprof
