+++ poseidon/bn254.go
@@ -28,7 +28,7 @@ type BN254State = [BN254_SPONGE_WIDTH]frontend.Variable
 type BN254HashOut = frontend.Variable
 
 func NewBN254Chip(api frontend.API) *BN254Chip {
-	return &BN254Chip{api: api, gl: *gl.New(api)}
+	return &BN254Chip{api: api, gl: *gl.NewChip(api)}
 }
 
 func (c *BN254Chip) Poseidon(state BN254State) BN254State {
