--- types/common_data_test.go
@@ -1,9 +1,0 @@
-package types
-
-import (
-	"testing"
-)
-
-func TestReadCommonCircuitData(t *testing.T) {
-	ReadCommonCircuitData("../testdata/decode_block/common_circuit_data.json")
-}
