+++ types/plonk.go
@@ -1,4 +1,4 @@
-package variables
+package types
 
 import gl "github.com/succinctlabs/gnark-plonky2-verifier/goldilocks"
