+++ verifier/data/decode_block/proof_with_public_inputs.json
