+++ verifier/data/decode_block/verifier_only_circuit_data.json
