+++ verifier/data/step/proof_with_public_inputs.json
