+++ verifier/data/step/verifier_only_circuit_data.json
